[TOA-35] ViewModels added to events

+++ app/src/main/java/pl/droidsonrioids/toast/data/api/EventService.kt
@@ -13,6 +13,6 @@ interface EventService {
     fun getEvents(): Single<EventsResponse>
 
     @GET("events/{id}")
-    fun getEvent(@Path("id") id: Int): Single<EventDetailsResponse>
+    fun getEvent(@Path("id") id: Long): Single<EventDetailsResponse>
 
 }
